requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
llama-cpp-python==0.2.11
sentence-transformers==2.2.2
chromadb==0.4.15
PyMuPDF==1.23.8
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3
torch>=2.0.0 